# [Deepin-Kernel-SIG] [linux 6.6-y] [Upstream] mm: improve performance of accounted kernel memory allocations